Version release: 18.1.3
Fix cache lock
The issue here is that when a fiber starts - the first thing it does is check whether it is cancelled. Therefore even if a fiber is evaluating something like

```scala
F.uncancelable { _ => F.never }
```

It will _not_ execute this if the fiber is cancelled before it actually runs.

Because of this the previous implementation of the cache had an issue where the foreground fiber would schedule the background fiber, get cancelled and therefore cancel the background fiber before the background fiber gets run. Therefore the background fiber would never start, and as a result - never clean up the cache entry for that key as intended. Instead the key would remain in the `Writing` state forever.

This changes the cache so that the cache cleanup no longer happens in the background fiber, to ensure we don't introduce any unintended cancelable gaps.